+++ src/main/java/com/tenten/linkhub/domain/space/controller/dto/space/SpacesFindByQueryApiResponses.java
@@ -1,15 +1,15 @@
 package com.tenten.linkhub.domain.space.controller.dto.space;
 
 import com.tenten.linkhub.domain.space.service.dto.SpacesFindByQueryResponses;
+import com.tenten.linkhub.global.PageMetaData;
 import org.springframework.data.domain.Slice;
 
 import java.util.List;
 
 public record SpacesFindByQueryApiResponses(
         List<SpacesFindByQueryApiResponse> responses,
-        Boolean hasNext,
-        Integer pageSize,
-        Integer pageNumber) {
+        PageMetaData metaData
+) {
 
     public static SpacesFindByQueryApiResponses from(SpacesFindByQueryResponses responses) {
         Slice<SpacesFindByQueryApiResponse> mapResponses = responses.responses()
@@ -22,11 +22,14 @@ public record SpacesFindByQueryApiResponses(
                         r.scrapCount(),
                         r.favoriteCount()));
 
-        return new SpacesFindByQueryApiResponses(
-                mapResponses.getContent(),
+        PageMetaData pageMetaData = new PageMetaData(
                 mapResponses.hasNext(),
                 mapResponses.getSize(),
                 mapResponses.getNumber());
+
+        return new SpacesFindByQueryApiResponses(
+                mapResponses.getContent(),
+                pageMetaData);
     }
 
 }
