+++ src/main/java/com/tenten/linkhub/domain/space/repository/space/DefaultSpaceRepository.java
@@ -2,16 +2,17 @@ package com.tenten.linkhub.domain.space.repository.space;
 
 import com.tenten.linkhub.domain.space.model.space.Space;
 import com.tenten.linkhub.domain.space.repository.space.dto.QueryCondition;
+import com.tenten.linkhub.domain.space.repository.space.query.SpaceQueryRepository;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class SpaceRepositoryImpl implements SpaceRepository {
+public class DefaultSpaceRepository implements SpaceRepository {
 
     private final SpaceJpaRepository spaceJpaRepository;
     private final SpaceQueryRepository spaceQueryRepository;
 
-    public SpaceRepositoryImpl(SpaceJpaRepository spaceJpaRepository, SpaceQueryRepository spaceQueryRepository) {
+    public DefaultSpaceRepository(SpaceJpaRepository spaceJpaRepository, SpaceQueryRepository spaceQueryRepository) {
         this.spaceJpaRepository = spaceJpaRepository;
         this.spaceQueryRepository = spaceQueryRepository;
     }
