+++ src/main/java/com/tenten/linkhub/global/PageMetaData.java
@@ -1,0 +1,8 @@
+package com.tenten.linkhub.global;
+
+public record PageMetaData(
+        Boolean hasNext,
+        Integer pageSize,
+        Integer pageNumber
+) {
+}
